feat(seamless limits): moved model into service/custodial for getting effective limits
