+++ packages/blockchain-wallet-v4-frontend/src/modals/Swap/EnterAmount/Checkout/index.tsx
@@ -15,11 +15,11 @@ import { AmountTextBox } from 'components/Exchange'
 import { FlyoutWrapper } from 'components/Flyout'
 import { convertBaseToStandard } from 'data/components/exchange/services'
 import { SwapAccountType, SwapBaseCounterTypes } from 'data/types'
+import { getEffectiveLimit } from 'services/custodial'
 import { formatTextAmount } from 'services/forms'
 import { media } from 'services/styles'
 
 import { OverLimitButton } from '../../../components'
-import { getEffectiveLimit } from '../../../model'
 import { StyledForm } from '../../components'
 import { Props as OwnProps, SuccessStateType } from '..'
 import { getMaxMin, incomingAmountNonZero, maximumAmount, minimumAmount } from './validation'
